Task 1: added exceptions to classes AirportFactory, PlaneFactory, Controller. User exceptions: EmptyAirportException, invalidCapacityException, InvalidTonnageException. System exceptions: FileNotFoundException, IOException
Task 2: added reading/writing to file using scanner (AirportFactory, PlaneFactory, Controller classes)
Task 3: added regexp for reading date from file (AirportFactory class)
Task 4: added factory classes (AirportFactory, PlaneFactory)
