+++ src/main/java/airport/entity/HeavyPlane.java
@@ -1,0 +1,15 @@
+package airport.entity;
+
+public class HeavyPlane extends Plane {
+
+
+    public HeavyPlane(String brand, int capacity, int tonnage, int range) {
+        super(brand, capacity, tonnage, range);
+    }
+
+    @Override
+    public void fly() {
+        System.out.println("Heavy plane is flying");
+    }
+
+}
