+++ src/main/java/airport/exceptions/EmptyAirportException.java
@@ -1,0 +1,7 @@
+package airport.exceptions;
+
+public class EmptyAirportException extends Exception {
+    public EmptyAirportException(String message){
+        super (message);
+    }
+}
