+++ src/main/java/airport/exceptions/InvalidCapacityException.java
@@ -1,0 +1,7 @@
+package airport.exceptions;
+
+public class InvalidCapacityException extends Exception {
+    public InvalidCapacityException(String message){
+        super (message);
+    }
+}
