+++ src/main/java/airport/exceptions/InvalidTonnageException.java
@@ -1,0 +1,7 @@
+package airport.exceptions;
+
+public class InvalidTonnageException extends Exception {
+    public InvalidTonnageException(String message){
+        super (message);
+    }
+}
