+++ src/main/java/airport/utility/PlaneType.java
@@ -1,0 +1,8 @@
+package airport.utility;
+
+public enum PlaneType {
+
+    HEAVY_PLANE,
+    REGIONAL_PLANE,
+    PRIVATE_PLANE;
+}
