--- src/main/java/lesson/lesson1/Task1/Main.java
@@ -1,15 +1,0 @@
-package lesson.lesson1.Task1;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-        //Task1.reverseSentence();
-        //Task1.splitSentence();
-        //Task1.spaceReplace();
-        //Task1.makeUpperCase();
-        //Task1.showTrimmedSentence();
-
-    }
-
-}
