--- src/main/java/lesson/lesson1/Task2/Main.java
@@ -1,11 +1,0 @@
-package lesson.lesson1.Task2;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-        Task2.showDayName();
-
-    }
-
-}
