--- src/main/java/lesson/lesson1/Task3/Main.java
@@ -1,11 +1,0 @@
-package lesson.lesson1.Task3;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-        Task3.drawRomb();
-
-    }
-
-}
