--- src/main/java/lesson/lesson1/Task5/Main.java
@@ -1,10 +1,0 @@
-package lesson.lesson1.Task5;
-
-public class Main {
-
-    public static void main(String[] args) {
-
-        Task5.guessTheName();
-    }
-
-}
