--- src/main/java/lesson/lesson2/Task1/entity/PrivatePlane.java
@@ -1,13 +1,0 @@
-package lesson.lesson2.Task1.entity;
-
-public class PrivatePlane extends Plane {
-
-    @Override
-    public void fly() {
-        System.out.println("Private plane is flying");
-    }
-
-    public PrivatePlane(String brand, int capacity, int tonnage, int range) {
-        super(brand, capacity, tonnage, range);
-    }
-}
