--- src/main/java/lesson/lesson2/Task1/entity/RegionalPlane.java
@@ -1,13 +1,0 @@
-package lesson.lesson2.Task1.entity;
-
-public class RegionalPlane extends Plane {
-
-    @Override
-    public void fly() {
-        System.out.println("Regional plane is flying");
-    }
-
-    public RegionalPlane(String brand, int capacity, int tonnage, int range) {
-        super(brand, capacity, tonnage, range);
-    }
-}
