--- src/main/java/lesson/lesson2/Task1/utility/PlaneType.java
@@ -1,8 +1,0 @@
-package lesson.lesson2.Task1.utility;
-
-public enum PlaneType {
-
-    HEAVY_PLANE,
-    REGIONAL_PLANE,
-    PRIVATE_PLANE;
-}
