--- src/main/java/lesson/lesson2/Task2/Student.java
@@ -1,49 +1,0 @@
-package lesson.lesson2.Task2;
-
-public class Student implements Comparable<Student> {
-
-    private String name;
-    private int age;
-    private String id;
-
-    public Student(String name, int age, String id) {
-        this.name = name;
-        this.age = age;
-        this.id = id;
-    }
-
-    @Override
-    public String toString() {
-        return name + " age: " + age + " id: " + id;
-    }
-
-    public int getAge() {
-        return age;
-    }
-
-    public void setAge(int age) {
-        this.age = age;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    @Override
-    public int compareTo(Student student) {
-        return (this.getAge() - student.getAge());
-    }
-}
-
