--- src/main/java/lesson/lesson3/airport/utility/PlaneType.java
@@ -1,8 +1,0 @@
-package lesson.lesson3.airport.utility;
-
-public enum PlaneType {
-
-    HEAVY_PLANE,
-    REGIONAL_PLANE,
-    PRIVATE_PLANE;
-}
